sections scenario updated
