+++ tests/e2e/main.scenario.js
@@ -1,7 +1,4 @@
 'use strict';
-describe('Main scenario', function() {
-  it('should load the main page', function() {
-    browser.get('/');
-    expect(browser.isElementPresent(By.css('body'))).toBe(true);
-  });
-});
+var sections = require('./sections.scenario');
+
+sections.run();
